Refactor SDK Manager to use Repository reader
